Flatten examples: move all .py files to examples/ root, delete subdirectories
- Move 20 example .py files from subdirectories to examples/
- Inline example_utils.py into us_bmode_linear_transducer.py
- Delete now-empty subdirectories and stale MATLAB file
- README.md renders below file list on GitHub (always visible)
