+++ examples/at_array_as_sensor.py
