+++ examples/at_array_as_source.py
