+++ examples/at_focused_annular_array_3D.py
