+++ examples/at_focused_bowl_3D.py
