+++ examples/at_focused_bowl_AS.py
