+++ examples/pr_2D_FFT_line_sensor.py
