+++ examples/pr_2D_TR_line_sensor.py
