+++ examples/pr_3D_FFT_planar_sensor.py
