+++ examples/pr_3D_TR_planar_sensor.py
