+++ examples/sd_directivity_modelling_2D.py
