+++ examples/sd_focussed_detector_2D.py
