+++ examples/sd_focussed_detector_3D.py
