+++ examples/us_defining_transducer.py
